DJANGO/Beauty/requirements.txt
django == 4.0.3
django-rest-framework
django-paranoid